[BIG update] Updated most of the existing packages & Substituted the old webpack configs with new ones.
Well, much of effort was put into this commit.

This project is based on anvaka/pm which is an old (but fantastic)
project came out in 2015. Both Webpack and React were in their early
versions. So were other packages used in this project.

So for this project, jumping from the old versions to the newest is
obviously troublesome. 总之我踩了好多坑.

The new webpack setup is just good but not perfect. It supports React
hot reload when in development mode and can output target files with
a smaller size (3.7MB compared with the old ~10MB) in production mode.

I am new to webpack configuration, maybe I can handle it more elegantly
in future.