Update packages & A new webpack setup & Remove useless files & Update info in package.json (#18)
* [refactor] Removed maco and rewrote all components in native React syntax

maco is a package written by anvaka to "avoid using javascript classes
when dealing with React". Now is 2021, both javascript and React have
moved forward a giant step, so maco is not needed any more.

I have removed maco from package.json and rewrote all involded code in
native React syntax.

* [typo] had missed one case that should have be internationalized, so fixed it.

* [refactor] Removed useless node templates.

* [BIG update] Updated most of the existing packages & Substituted the old webpack configs with new ones.

Well, much of effort was put into this commit.

This project is based on anvaka/pm which is an old (but fantastic)
project came out in 2015. Both Webpack and React were in their early
versions. So were other packages used in this project.

So for this project, jumping from the old versions to the newest is
obviously troublesome. 总之我踩了好多坑.

The new webpack setup is just good but not perfect. It supports React
hot reload when in development mode and can output target files with
a smaller size (3.7MB compared with the old ~10MB) in production mode.

I am new to webpack configuration, maybe I can handle it more elegantly
in future.

* [typo] Due to update of less, a/b should be (a/b) now.

* [refactor] Removed files not needed any more.

* [doc] Updated basic info in package.json

* [fix] output.publicPath in webpack.config.js is wrong, so fix it.

- publicPath: "/" --> relative to server
- publicPath: "" --> relative to HTML file

Co-authored-by: 唐烨男 <tangyenan@MacBook-M1-Pro.local>